Issue fixed on lensFacing
